requirement.txt
streamlit
PyPDF2
google.generativeai
python-dotenv
langchain
langchain_community
langchain_google_genai
